+++ blocklets/ai-studio/api/src/store/projects.ts
@@ -69,6 +69,7 @@ export const projectTemplates: (Project & {
         updatedBy: wallet.address,
         createdAt: '2023-09-30T12:23:04.603Z',
         updatedAt: '2023-09-30T12:23:04.603Z',
+        public: false,
       },
     ],
   },