+++ blocklets/ai-studio/migration/0.1.97.js
@@ -1,1 +1,0 @@
-require('../api/dist/migration/0.1.97');