+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 0.1.97 (2023-10-16)
+## 0.1.98 (2023-10-16)
 
 - fix: set default model for project
+
+## 0.1.97 (2023-10-16)
+
+- chore: export resource data
 
 ## 0.1.96 (2023-10-16)